Create route api
