+++ api/controllers/user.controller.js
@@ -1,0 +1,3 @@
+export const test = (req, res) => {
+  res.json({ message: "Api is running !..." });
+};
