--- api/controllers/user.controllers.js
@@ -1,3 +1,0 @@
-export const test = (req, res) => {
-  res.json({ message: "Api is running..." });
-};
