--- api/modules/user.modul.js
@@ -1,27 +1,0 @@
-import mongoose from "mongoose";
-
-const userSchema = new mongoose.Schema({
-    username: {
-        type: String,
-        required: true,
-        min: 3,
-        max: 20,
-        unique: true
-    },
-    email: {
-        type: String,
-        required: true,
-        unique: true
-    },
-    password: {
-        type: String,
-        required: true,
-        min: 6
-    },
-},({timestamps: true}));
-
-const User = mongoose.model("User", userSchema);
-
-
-export default User;
-
